% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_numbers}
\alias{format_numbers}
\title{Formats numbers for text output}
\usage{
format_numbers(x, capitalized = FALSE)
}
\arguments{
\item{x}{A numeric value.}

\item{capitalized}{A logical value indicating if the output should start with
a capital letter (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A character value with the formatted value from \code{x}.
}
\description{
Formats numbers for text output
}
