% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkl_get_normalized_peak_heights.R
\name{hkl_get_normalized_peak_heights}
\alias{hkl_get_normalized_peak_heights}
\title{Extracts normalized peak heights for a set of peaks from mid infrared spectra}
\usage{
hkl_get_normalized_peak_heights(x, ..., do_scale = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link[ir:ir_new_ir]{ir}}.}

\item{...}{Further arguments passed to
\code{pmird:::irp_content_klh_hodgkins_main}.}

\item{scale}{A logical value indicating if peak heights should be
z-transformed (\code{TRUE}) or not (\code{FALSE}), using
\code{\link{scale}}.}
}
\value{
A data frame with a row for each spectrum in \code{x} and a column for
each peak extracted by \code{pmird:::irp_content_klh_hodgkins_main}.
}
\description{
Extracts normalized peak heights for a set of peaks from mid infrared spectra
}
