% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkl_hodgkins_ggplot.R
\name{hkl_hodgkins_ggplot}
\alias{hkl_hodgkins_ggplot}
\title{Plots spectra with identified peaks as ggplot2 plots}
\source{
The code is modified from
\url{https://github.com/shodgkins/FTIRbaselines}.
}
\usage{
hkl_hodgkins_ggplot(x)
}
\arguments{
\item{x}{A data frame as returned by
\code{irpeat:::irp_content_klh_hodgkins_prepare}.}
}
\value{
A list of \code{ggplot2-object} with a plot for each spectrum in
\code{x}.
}
\description{
Plots spectra with identified peaks as ggplot2 plots
}
\details{
\code{hkl_hodgkins_ggplot} is a modified part of the code from
\url{https://github.com/shodgkins/FTIRbaselines}. It extract a defined set of
peaks from mid infrared spectra and creates a plot of these peaks as
\code{\link[ggplot2:ggplot2]{ggplot2-object}}.
}
