% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hklmirs-package.R
\docType{package}
\name{hklmirs-package}
\alias{hklmirs}
\alias{hklmirs-package}
\title{hklmirs: Reproducible Research Compendium for "Improving Models to Predict Holocellulose and Klason Lignin Contents for Peat Soil Organic Matter with Mid Infrared Spectra" and "Predicting Absolute Holocellulose and Klason Lignin Contents for Peat Remains Challenging".}
\description{
Reproducible Research Compendium for "Improving Models to Predict Holocellulose and Klason Lignin Contents for Peat Soil Organic Matter with Mid Infrared Spectra" and "Predicting Absolute Holocellulose and Klason Lignin Contents for Peat Remains Challenging". In these two manuscripts, a concrete strategy to improve original models developed by Hodgkins et al. (2018) to predict organic matter holocellulose and Klason lignin contents from mid infrared spectra is proposed and the results from Hodgkins et al. (2018) are replicated.
}
\author{
\strong{Maintainer}: Henning Teickner \email{h_teic01@uni-muenster.de} (\href{https://orcid.org/0000-0002-3993-1182}{ORCID})

Authors:
\itemize{
  \item Klaus-Holger Knorr \email{kh.knorrr@wwu.de} (\href{https://orcid.org/0000-0003-4175-0214}{ORCID})
}

}
\keyword{internal}
